ai_bot/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
cloudpickle==1.2.2
discord.py==1.3.1
future==0.18.2
gym==0.16.0
idna==2.9
idna-ssl==1.1.0
multidict==4.7.5
numpy==1.18.1
Pillow==7.0.0
pyglet==1.5.0
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
typing-extensions==3.7.4.1
websockets==8.1
yarl==1.4.2
